Fix printf("%x"...) for 64-bit values

+++ printf.c
@@ -15,7 +15,7 @@ printint(int fd, int64 xx, int base, int sgn)
   static char digits[] = "0123456789ABCDEF";
   char buf[16];
   int i, neg;
-  uint x;
+  uint64 x;
 
   neg = 0;
   if(sgn && xx < 0){
